Publish built package (#781)
* Introduce publish config property

* Introduce custom publish command configuration

* Prepare publish build step

* introduce publish buiuld step

* introduce publish build item to order build steps

+++ deployment/src/main/java/io/quarkiverse/quinoa/deployment/QuinoaProcessor.java
@@ -34,6 +34,7 @@ import io.quarkiverse.quinoa.deployment.framework.FrameworkType;
 import io.quarkiverse.quinoa.deployment.items.BuiltResourcesBuildItem;
 import io.quarkiverse.quinoa.deployment.items.ConfiguredQuinoaBuildItem;
 import io.quarkiverse.quinoa.deployment.items.InstalledPackageManagerBuildItem;
+import io.quarkiverse.quinoa.deployment.items.PublishedPackageBuildItem;
 import io.quarkiverse.quinoa.deployment.items.TargetDirBuildItem;
 import io.quarkiverse.quinoa.deployment.packagemanager.PackageManagerInstall;
 import io.quarkiverse.quinoa.deployment.packagemanager.PackageManagerRunner;
@@ -41,6 +42,7 @@ import io.quarkiverse.quinoa.deployment.packagemanager.types.PackageManagerType;
 import io.quarkus.deployment.IsDevelopment;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.annotations.Consume;
 import io.quarkus.deployment.annotations.Record;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 import io.quarkus.deployment.builditem.HotDeploymentWatchedFileBuildItem;
@@ -186,6 +188,12 @@ public class QuinoaProcessor {
                     Set.of("quarkus.quinoa.build-dir"));
         }
 
+        // doesn't make sense to copy from ui build dir to quinoa target dir
+        // in case of `just-build` option enabled
+        if (configuredQuinoa.resolvedConfig().justBuild()) {
+            return new TargetDirBuildItem(buildDir);
+        }
+
         final Path targetBuildDir = initializeTargetDirectory(outputTarget).resolve(TARGET_BUILD_DIR_NAME);
         FileUtil.deleteDirectory(targetBuildDir);
         try {
@@ -201,10 +209,37 @@ public class QuinoaProcessor {
     }
 
     @BuildStep
-    public BuiltResourcesBuildItem prepareBuiltResources(Optional<TargetDirBuildItem> targetDir) throws IOException {
+    public PublishedPackageBuildItem publishBuiltPackage(
+            ConfiguredQuinoaBuildItem configuredQuinoa,
+            InstalledPackageManagerBuildItem installedPackageManager,
+            Optional<TargetDirBuildItem> targetDir) {
+        if (configuredQuinoa == null || !configuredQuinoa.resolvedConfig().publish()) {
+            return new PublishedPackageBuildItem(true);
+        }
+
+        if (targetDir.isEmpty()) {
+            LOG.warn("Target dir is empty - unable to run publish command.");
+            return new PublishedPackageBuildItem(true);
+        }
+
+        final PackageManagerRunner packageManagerRunner = installedPackageManager.getPackageManager();
+        packageManagerRunner.publish();
+        return new PublishedPackageBuildItem(false);
+    }
+
+    @BuildStep
+    @Consume(PublishedPackageBuildItem.class) // just to order build steps
+    public BuiltResourcesBuildItem prepareBuiltResources(
+            ConfiguredQuinoaBuildItem configuredQuinoa,
+            Optional<TargetDirBuildItem> targetDir) throws IOException {
         if (targetDir.isEmpty()) {
             return null;
         }
+        if (configuredQuinoa != null && configuredQuinoa.resolvedConfig().justBuild()) {
+            // no need to configure vertx static resources when `just-build` activated
+            return null;
+        }
+
         return new BuiltResourcesBuildItem(lookupBuiltResources(targetDir.get().getBuildDirectory()));
     }
